Extract utils libraries for core, collection (#648)
This avoids mixing implementation into umbrella libraries.
