+++ test/collection/all_tests.dart
@@ -15,7 +15,6 @@
 library quiver.collection.all_tests;
 
 import 'bimap_test.dart' as bimap;
-import 'collection_test.dart' as collection;
 import 'delegates/iterable_test.dart' as delegate_iterable;
 import 'delegates/list_test.dart' as delegate_list;
 import 'delegates/map_test.dart' as delegate_map;
@@ -24,10 +23,10 @@ import 'delegates/set_test.dart' as delegate_set;
 import 'lru_map_test.dart' as lru_map;
 import 'multimap_test.dart' as multimap;
 import 'treeset_test.dart' as treeset;
+import 'utils_test.dart' as utils;
 
 void main() {
   bimap.main();
-  collection.main();
   delegate_iterable.main();
   delegate_list.main();
   delegate_map.main();
@@ -36,4 +35,5 @@ void main() {
   lru_map.main();
   multimap.main();
   treeset.main();
+  utils.main();
 }
